worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-auth==2.23.4
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
python-docx==1.1.0
PyPDF2==3.0.1
markdown==3.5.1
zipfile-deflate64==0.2.0
google-generativeai==0.3.2
openai==1.3.7
numpy==1.24.3
scikit-learn==1.3.0
